[BACKEND] Fix scans with duplicated elements
Say we run the following with 128 threads:
```python
tmp0 = tl.full([64], 1, tl.int32)
tmp1 = tl.cumsum(tmp, 0)
```

I see a scan is performed across all 128 threads, giving a maximum value
larger than the sum of `tmp0`. This is observable if you broadcast the
result to a larger shape, or if you have device asserts which observe the output.

This fixes the issue by treating duplicated elements in the scan axis as a new
axis along which we do further parallel reductions. This is admittedly
inefficient, but at least fixes the correctness issues.