[OPTIMIZER] Allow rematerializing through scf.for
Currently values inside a for loop can be rematerialized, but the rematerialized
values returned by the for loop won't be used outside the loop and layout
conversions outside a loop won't trigger rematerialization with the loop.

Strangely enough the rewrite code does support this, but the slice creation
doesn't handle it so it wasn't reachable.