[OPTIMIZER] Improve select masked load pattern
This ports the select masked load pattern to tablegen and adds an extra
condition that the load only has one use so we don't duplicate loads.

I also expand the pattern to work with non-splat masks.

Finally, I remove `CombineBroadcastConstantPattern` as there's already a
constant folder for `BroadcastOp` which covers this pattern.